Separate polygon into header and cpp
